Update match API (#261)
This PR updates the match API to ensure valid entries and enforce
admin-only authorization.

+++ app/routers/matches.py
@@ -6,6 +6,8 @@ from fastapi.exceptions import HTTPException
 from pydantic.v1 import root_validator
 from sqlmodel import Session, select
 
+from app.database import flights as db_flights
+from app.database import teams as db_teams
 from app.dependencies import get_current_active_user, get_sql_db_session
 from app.models.base import APLGLBaseModel
 from app.models.flight import Flight
@@ -105,10 +107,41 @@ async def create_match(
     current_user: User = Depends(get_current_active_user),
     match: MatchCreate,
 ):
+    if not current_user.is_admin:
+        raise HTTPException(
+            status_code=HTTPStatus.UNAUTHORIZED,
+            detail=f"User '{current_user.name}' not authorized to create matches",
+        )
+
+    flight = db_flights.get_by_id(session=session, flight_id=match.flight_id)
+    if flight is None:
+        raise HTTPException(
+            status_code=HTTPStatus.NOT_FOUND,
+            detail=f"Unable to find flight with id: {match.flight_id}",
+        )
+    if match.week > flight.weeks:
+        raise HTTPException(
+            status_code=HTTPStatus.BAD_REQUEST,
+            detail=f"Unable to create match on week '{match.week}' for flight with '{flight.weeks}' weeks",
+        )
+    home_team = db_teams.get_by_id(session=session, team_id=match.home_team_id)
+    if home_team is None:
+        raise HTTPException(
+            status_code=HTTPStatus.NOT_FOUND,
+            detail=f"Unable to find home team with id: {match.home_team_id}",
+        )
+    away_team = db_teams.get_by_id(session=session, team_id=match.away_team_id)
+    if away_team is None:
+        raise HTTPException(
+            status_code=HTTPStatus.NOT_FOUND,
+            detail=f"Unable to find away team with id: {match.away_team_id}",
+        )
+
     match_db = Match.model_validate(match)
     session.add(match_db)
     session.commit()
     session.refresh(match_db)
+
     return match_db
 
 
@@ -128,15 +161,24 @@ async def update_match(
     match_id: int,
     match: MatchUpdate,
 ):
+    if not current_user.is_admin:
+        raise HTTPException(
+            status_code=HTTPStatus.UNAUTHORIZED,
+            detail=f"User '{current_user.name}' not authorized to update matches",
+        )
+
     match_db = session.get(Match, match_id)
     if not match_db:
         raise HTTPException(status_code=404, detail="Match not found")
+
     match_data = match.model_dump(exclude_unset=True)
     for key, value in match_data.items():
         setattr(match_db, key, value)
+
     session.add(match_db)
     session.commit()
     session.refresh(match_db)
+
     return match_db
 
 
@@ -147,12 +189,21 @@ async def delete_match(
     current_user: User = Depends(get_current_active_user),
     match_id: int,
 ):
+    if not current_user.is_admin:
+        raise HTTPException(
+            status_code=HTTPStatus.UNAUTHORIZED,
+            detail=f"User '{current_user.name}' not authorized to delete matches",
+        )
+
     match_db = session.get(Match, match_id)
     if not match_db:
         raise HTTPException(status_code=404, detail="Match not found")
+
     session.delete(match_db)
     session.commit()
+
     # TODO: Delete related resources (match-round-links)
+
     return {"ok": True}
 
 
@@ -164,6 +215,11 @@ async def post_match_rounds(
     match_input: MatchInput,
 ):
     # TODO: Check user credentials
+    if not (current_user.is_admin or current_user.edit_flights):
+        raise HTTPException(
+            status_code=HTTPStatus.UNAUTHORIZED,
+            detail=f"User '{current_user.name}' not authorized to enter match scores",
+        )
     ahs = APLHandicapSystem()
 
     match_db = session.get(Match, match_input.match_id)
